unifying equality: move to own module
