+++ rust/ares/src/lib.rs
@@ -15,6 +15,7 @@ pub mod serf;
 pub mod persist;
 pub mod serialization;
 pub mod trace;
+pub mod unifying_equality;
 
 /** Introduce useful functions for debugging
  *
